company테이블 DTO클래스,mapper,service 수정 3, 톰캣 연결되는데 companydetail 클래스를 갑자기 인식을 못함
