+++ src/main/java/org/scoula/config/RootConfig.java
@@ -19,9 +19,8 @@ import javax.sql.DataSource;
 @Configuration
 
 @ComponentScan(basePackages = "org.scoula")
-@MapperScan(basePackages = "org.scoula.mapper")
+@MapperScan(basePackages = {"org.scoula.mapper", "org.scoula.user.mapper"})
 @PropertySource({"classpath:/application.properties"})
-@MapperScan(basePackages = "org.scoula.user.mapper")
 public class RootConfig {
     @Value("${jdbc.driver}")
     String driver;
@@ -56,8 +55,6 @@ public class RootConfig {
         SqlSessionFactoryBean sqlSessionFactory = new SqlSessionFactoryBean();
         sqlSessionFactory.setConfigLocation(applicationContext.getResource("classpath:/mybatis-config.xml"));
         sqlSessionFactory.setDataSource(dataSource());
-
-        sqlSessionFactory.setMapperLocations(applicationContext.getResources("classpath:/mapper/*.xml"));
         return (SqlSessionFactory) sqlSessionFactory.getObject();
     }
