+++ src/main/java/org/scoula/dto/PostDTO.java
@@ -1,4 +1,4 @@
-package org.scoula.DTO;
+package org.scoula.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
